% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gChain.R
\docType{methods}
\name{c}
\alias{c}
\title{c}
\usage{
\S4method{c}{gChain}(x, ..., recursive = FALSE)
}
\description{
concatenate gChains, i.e. concatenate their links
they must have the same domain and range genome
}

