% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gChain.R
\name{cgChain}
\alias{cgChain}
\title{cgChain}
\usage{
cgChain(cigar, sn = NULL, verbose = TRUE)
}
\description{
cgChain (i.e. "CIGAR")

Processes a pairwise alignment from read to genomic coordinates inputted as a vector of CIGAR strings
representing edit operations from subject to pattern.
ie an insertion corresponds a gap in the subject and deletion corresponds to a gap in the pattern
returns a chain that maps subject to pattern (e.g. genome space to read space)

Also, "cigar" can be a GRanges or GappedAlignment object with $cigar field where the ranges specify flanking subject coordinates
of the aligned pattern (e.g. read).  As by convention, subject coordinates are assumed to NOT contain flanking soft clips.
Internal soft clips (if any exist - this would be unconventional use) are treated as insertions.
If seqnames ("sn") input not provided then the read seqnames will be the $qname field + IsFirstMateRead part of read flag.

Returns a gChain mapping read space to genome space.
cigar --- ## can be character vector or GRanges GAlignment with $cigar and $qname string field)
}
\author{
Marcin Imielinski
}
