% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gChain.R
\name{cn}
\alias{cn}
\title{cn}
\usage{
cn(x, ...)
}
\description{
A method to extract copy number from a chain mapping genome x to y

Returns a GRanges in genome x specifying the number of copies of every interval
in x that are contained in y

as "breaks" only goes in the "forward" direction
}
