% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gChain.R
\docType{methods}
\name{gMultiply}
\alias{gMultiply}
\title{gMultiply}
\usage{
gMultiply(e1, e2, pintersect = NA)

\S4method{*}{gChain,gChain}(e1, e2)
}
\description{
gChain multiply
gchain::multiply

if e2 lifts from genome A to genome B
and e1 lifts from genome B to genome C
then e1 * e2 = lifts from genome A to genome C
}

