% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gChain.R
\name{gSubset}
\alias{gSubset}
\title{gSubset}
\usage{
gSubset(gc, xnames = NULL, ynames = NULL, x.or.y = FALSE)
}
\arguments{
\item{gc}{gChain to subset}

\item{xnames}{character vector that has seqnames of links(gc)$x to keep}

\item{ynames}{character vector that has seqnames of links(gc)$y to keep}

\item{x.or.y}{Links with hits in xnames OR ynames included. Default FALSE (i.e. AND)}
}
\value{
subsetted gChain
}
\description{
Subset a gChain

Make smaller gChain that contains only the seqnames
given by xnames and/or ynames.
}

