% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gChain.R
\name{gUnique}
\alias{gUnique}
\title{gUnique}
\usage{
gUnique(gc)
}
\arguments{
\item{gc}{gChain object to deduplicate}
}
\value{
Deduplicated gChain
}
\description{
remove duplicate x to y mappings

Will select only the longest, and most "leftward" in genomic coordaintes
(in that order) entry of each unique x to y pair, where
uniqueness is determined by seqnames and strand, but not by position.
This could occur, for instance, if a read has multiple mappings to a
genomic contig or multiple mappings within one chromosome. Occasionally,
one might want to take only the longest match, and discard secondary alignments.
}
\note{
This is a bit of a hack, as there are better ways to remove dupes before this step
}

