% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gChain.R
\name{initialize}
\alias{initialize}
\title{initialize}
\usage{
\S4method{initialize}{gChain}(
  .Object,
  x = NULL,
  y = NULL,
  pad.left = 0,
  pad.right = 0,
  scale = NULL,
  val = data.frame()
)
}
\description{
gchain::initialize
gchain initialize

if only x (or y) GRanges is given will make "identity chain"
if x (or y) is given as a seqinfo, then will make "empty chain"
}
