% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gChain.R
\name{maChain}
\alias{maChain}
\title{maChain}
\usage{
maChain(grl = NULL, pali, pad = 0, trim = T, trim.thresh = 0)
}
\description{
multiple sequence alignment chain

Takes input GRanges List of interval coordinates corresponding to sequences
contributing to alignments in pali. 'pali' is a list of alignment matrices
(containing letters from the original sequences and '-', '.' for gaps)

'grl' and 'pali', have the same names (corresponding to the alignment, e.g. a protein domain)
Also, this property holds:
sapply(pali, nrow) = sapply(grl, length)

trim will remove all alignment columns that have only gaps

alternatively .. grl can be an XStringSet object with a "names" attribute containing
Biostrings (eg AA or DNA) and gap ("-") characters.  Can also be a list of XStringSet objects.

if grl is NULL, then it is assumed that the first (last) non gapped character in each MSA alignment entry
represents the first (last) item of the sequence
}
\author{
Marcin Imielinski
}
