% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gChain.R
\name{paChain}
\alias{paChain}
\title{paChain}
\usage{
paChain(seq1, seq2, sn1 = NULL, sn2 = NULL, gr1 = NULL, gr2 = NULL,
  pa = NULL, verbose = T, sl1 = NULL, sl2 = NULL, score.thresh = NULL,
  both.strands = F, keep.best = T, assume.dna = T, mc.cores = 1,
  mc.chunks = mc.cores, pintersect = FALSE, ...)
}
\description{
make chain from pairwiseAlignment object

each BSStringSet must have named sequences (otherwise names will be computed
from (perfectly unique) sequences.  Each name can only map to a single sequence

NOTE: if XStringSet objects are not provided, then the input is assumed to be DNA
(if assume.dna = T), otherwise AA (if assume.dna = F)

Pre-computed pa object can be also provided, but then the sequences from which the
pa object was computed should be given as well.

returns gChain with seqinfos corresponding to unique seqnames in seq1 and seq2
and the interval pairs mapping those two sequences
}
\author{
Marcin Imielinski
}

