% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gChain.R
\name{squeeze}
\alias{squeeze}
\title{squeeze}
\usage{
squeeze(x, gap = 0)
}
\arguments{
\item{x}{IRanges object}
}
\value{
squeezed IRanges object
}
\description{
"squeezes" pile of IRanges so that width(squeeze(ir)) = width(ir), end(ir[length(ir)]) = sum(width(ir)) + 1
start(squeeze(ir)[k]) = end(squeeze(ir)[k-1])+1 for all k>1, and start(squeeze(ir))[1] = 1

optional extra input "gap" yields following output: width(squeeze(ir, gap)) = width(ir), end(ir[length(ir)]) = sum(width(ir)) + gap*(length(ir)-1) + 1
start(squeeze(ir, gap)[k]) = end(squeeze(ir, gap)[k-1])+1+gap for all k>1, and start(squeeze(ir, gap))[1] = 1
}
