% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gChain.R
\docType{methods}
\name{values}
\alias{values}
\alias{values}
\alias{values}
\title{values}
\usage{
\S4method{values}{gChain}(x)

\S4method{scale}{gChain}(x)

pads(.Object, ...)
}
\description{
Print meta data of gChain

Get scale of gChain

Get pads of gChain
}
