% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gChain.R
\name{breaks}
\alias{breaks}
\title{breaks}
\usage{
breaks(x, ...)
}
\description{
Method to extract "breaks" from a gChain mapping genome x to y
i.e. pairs of positions that are contiguous in y but were non-contiguous in x

Returns a GRangesList in genome x containing the location of pairs of positions
that are contiguous in y that were not contiguous in x.

Breaks are only defined in one direction (ie x to y)
(if you want to do reverse then just "transpose" the chain, or use rev = TRUE)
}
