% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gChain.R
\name{txChain}
\alias{txChain}
\title{txChain}
\usage{
txChain(grl, txname = NULL, translate = F, exonFrame.field = "exon_frame",
  val = NULL)
}
\description{
Transcript chain

Takes a GRangesList on genomic coordinates (ie exons comprising transcripts)
and creates a gChain representing mapping onto transcript (or translated protein) coordinates
}
\author{
Marcin Imielinski
}
